Remove superfluous grid line for log scale.
The code removed in this change added an additional entry to the list of
labels, causing the top of the graph to be extended by self.logBase. As
the corresponding labels have already been drawn, the result is a
graph with n+1 gridlines for n labels. Removing this addition resolves
the issue.

Fixes #235.